// backend/src/application/use-cases/interfaces/company/i-create-employee-use-case.ts
import { Employee } from "../../../../domain/entities/employee";



export interface ICreateEmployeeUseCase{
    execute(employee: {
    name: string;
    email: string;
    phone: string;
    dob: Date;
    joiningDate: Date;
    position: string;
    departmentId: string;
    status: "active" | "inactive" | "suspended";
  }):Promise<Employee>;

} 